// src/common/enums/transaction-state.enum.ts
// src/common/enums/transaction-state.enum.ts

export enum TransactionState {
  CREATED = 'CREATED',
  ROUTE_SELECTED = 'ROUTE_SELECTED',
  AUTH_INITIATED = 'AUTH_INITIATED',
  AUTHORISED = 'AUTHORISED',
  AUTH_FAILED = 'AUTH_FAILED',
  AUTH_TIMEOUT = 'AUTH_TIMEOUT',
  AUTH_EXPIRED = 'AUTH_EXPIRED',
  CAPTURE_INITIATED = 'CAPTURE_INITIATED',
  CAPTURED = 'CAPTURED',
  PARTIALLY_CAPTURED = 'PARTIALLY_CAPTURED',
  CAPTURE_FAILED = 'CAPTURE_FAILED',
  VOID_INITIATED = 'VOID_INITIATED',
  VOIDED = 'VOIDED',
  REFUND_INITIATED = 'REFUND_INITIATED',
  REFUNDED = 'REFUNDED',
  PARTIALLY_REFUNDED = 'PARTIALLY_REFUNDED',
  REFUND_FAILED = 'REFUND_FAILED',
  SETTLED = 'SETTLED',
  DISPUTE_OPENED = 'DISPUTE_OPENED',
  DISPUTE_RESOLVED = 'DISPUTE_RESOLVED',
  ABANDONED = 'ABANDONED',
  ROUTE_FAILED = 'ROUTE_FAILED',
  FAILED = 'FAILED',
}